chore(ui/dataLoaders): Move dataloaders actions and reducers to dataLoaders dir
