+++ ui/src/onboarding/components/configureStep/ConfigureDataSourceSwitcher.tsx
@@ -15,7 +15,7 @@ import {
   addConfigValue,
   removeConfigValue,
   setConfigArrayValue,
-} from 'src/onboarding/actions/dataLoaders'
+} from 'src/dataLoaders/actions/dataLoaders'
 
 // Types
 import {TelegrafPlugin, DataLoaderType, Substep} from 'src/types/v2/dataLoaders'
