+++ ui/src/onboarding/components/configureStep/Scraping.tsx
@@ -14,8 +14,8 @@ import {
   setScraperTargetBucket,
   setScraperTargetURL,
   saveScraperTarget,
-} from 'src/onboarding/actions/dataLoaders'
-import {setBucketInfo} from 'src/onboarding/actions/steps'
+} from 'src/dataLoaders/actions/dataLoaders'
+import {setBucketInfo} from 'src/dataLoaders/actions/steps'
 
 // Types
 import {Bucket} from 'src/api'
