+++ ui/src/onboarding/components/configureStep/lineProtocol/LineProtocolTabs.tsx
@@ -19,7 +19,7 @@ import {
   setActiveLPTab,
   writeLineProtocolAction,
   setPrecision,
-} from 'src/onboarding/actions/dataLoaders'
+} from 'src/dataLoaders/actions/dataLoaders'
 
 import {AppState} from 'src/types/v2/index'
