+++ ui/src/onboarding/components/configureStep/lineProtocol/TabBody.tsx
@@ -11,7 +11,7 @@ import {
 import DragAndDrop from 'src/shared/components/DragAndDrop'
 import TextArea from 'src/clockface/components/inputs/TextArea'
 import {LineProtocolTab} from 'src/types/v2/dataLoaders'
-import {setLineProtocolBody} from 'src/onboarding/actions/dataLoaders'
+import {setLineProtocolBody} from 'src/dataLoaders/actions/dataLoaders'
 import {WritePrecision} from 'src/api'
 
 interface Props {
