+++ ui/src/onboarding/components/configureStep/streaming/PluginConfigForm.tsx
@@ -13,7 +13,7 @@ import {
   addConfigValue,
   removeConfigValue,
   setConfigArrayValue,
-} from 'src/onboarding/actions/dataLoaders'
+} from 'src/dataLoaders/actions/dataLoaders'
 
 // Types
 import {TelegrafPlugin, ConfigFields} from 'src/types/v2/dataLoaders'
