+++ ui/src/onboarding/components/configureStep/streaming/PluginConfigSwitcher.tsx
@@ -15,7 +15,7 @@ import {
   addConfigValue,
   removeConfigValue,
   setConfigArrayValue,
-} from 'src/onboarding/actions/dataLoaders'
+} from 'src/dataLoaders/actions/dataLoaders'
 
 // Types
 import {TelegrafPlugin, Substep} from 'src/types/v2/dataLoaders'
